+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/exporter/builders/EventTelemetryBuilder.java
@@ -19,59 +19,32 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.exporter.models.builders;
+package com.microsoft.applicationinsights.agent.internal.exporter.builders;
 
 import static com.microsoft.applicationinsights.agent.internal.common.TelemetryTruncation.truncateTelemetry;
 
 import com.microsoft.applicationinsights.agent.internal.common.Strings;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.RequestData;
-import com.microsoft.applicationinsights.agent.internal.exporter.utils.SanitizationHelper;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.TelemetryEventData;
 import java.util.HashMap;
 import java.util.Map;
 
-public final class RequestTelemetryBuilder extends AbstractTelemetryBuilder {
+public final class EventTelemetryBuilder extends AbstractTelemetryBuilder {
 
-  private static final int MAX_SOURCE_LENGTH = 1024;
-  private static final int MAX_RESPONSE_CODE_LENGTH = 1024;
+  private static final int MAX_EVENT_NAME_LENGTH = 512;
 
-  private final RequestData data;
+  private final TelemetryEventData data;
 
-  public static RequestTelemetryBuilder create() {
-    return new RequestTelemetryBuilder(new RequestData());
+  public static EventTelemetryBuilder create() {
+    return new EventTelemetryBuilder(new TelemetryEventData());
   }
 
-  private RequestTelemetryBuilder(RequestData data) {
-    super(data, "Request", "RequestData");
+  private EventTelemetryBuilder(TelemetryEventData data) {
+    super(data, "Event", "EventData");
     this.data = data;
   }
 
-  public void setId(String id) {
-    data.setId(truncateTelemetry(id, SanitizationHelper.MAX_ID_LENGTH, "Request.id"));
-  }
-
   public void setName(String name) {
-    data.setName(truncateTelemetry(name, SanitizationHelper.MAX_NAME_LENGTH, "Request.name"));
-  }
-
-  public void setDuration(String duration) {
-    data.setDuration(duration);
-  }
-
-  public void setSuccess(boolean success) {
-    data.setSuccess(success);
-  }
-
-  public void setResponseCode(String responseCode) {
-    data.setResponseCode(
-        truncateTelemetry(responseCode, MAX_RESPONSE_CODE_LENGTH, "Request.responseCode"));
-  }
-
-  public void setSource(String source) {
-    data.setSource(truncateTelemetry(source, MAX_SOURCE_LENGTH, "Request.source"));
-  }
-
-  public void setUrl(String url) {
-    data.setUrl(truncateTelemetry(url, SanitizationHelper.MAX_URL_LENGTH, "Request.url"));
+    data.setName(truncateTelemetry(name, MAX_EVENT_NAME_LENGTH, "Event.name"));
   }
 
   public void addMeasurement(String key, Double value) {
