+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/exporter/builders/MetricPointBuilder.java
@@ -19,7 +19,7 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.exporter.models.builders;
+package com.microsoft.applicationinsights.agent.internal.exporter.builders;
 
 import static com.microsoft.applicationinsights.agent.internal.common.TelemetryTruncation.truncateTelemetry;
