+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/exporter/builders/RemoteDependencyTelemetryBuilder.java
@@ -19,32 +19,66 @@
  * DEALINGS IN THE SOFTWARE.
  */
 
-package com.microsoft.applicationinsights.agent.internal.exporter.models.builders;
+package com.microsoft.applicationinsights.agent.internal.exporter.builders;
 
 import static com.microsoft.applicationinsights.agent.internal.common.TelemetryTruncation.truncateTelemetry;
 
 import com.microsoft.applicationinsights.agent.internal.common.Strings;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.TelemetryEventData;
+import com.microsoft.applicationinsights.agent.internal.exporter.models.RemoteDependencyData;
+import com.microsoft.applicationinsights.agent.internal.exporter.utils.SanitizationHelper;
 import java.util.HashMap;
 import java.util.Map;
 
-public final class EventTelemetryBuilder extends AbstractTelemetryBuilder {
+public final class RemoteDependencyTelemetryBuilder extends AbstractTelemetryBuilder {
 
-  private static final int MAX_EVENT_NAME_LENGTH = 512;
+  private static final int MAX_DATA_LENGTH = 8192;
+  private static final int MAX_RESULT_CODE_LENGTH = 1024;
+  private static final int MAX_DEPENDENCY_TYPE_LENGTH = 1024;
+  private static final int MAX_TARGET_NAME_LENGTH = 1024;
 
-  private final TelemetryEventData data;
+  private final RemoteDependencyData data;
 
-  public static EventTelemetryBuilder create() {
-    return new EventTelemetryBuilder(new TelemetryEventData());
+  public static RemoteDependencyTelemetryBuilder create() {
+    return new RemoteDependencyTelemetryBuilder(new RemoteDependencyData());
   }
 
-  private EventTelemetryBuilder(TelemetryEventData data) {
-    super(data, "Event", "EventData");
+  private RemoteDependencyTelemetryBuilder(RemoteDependencyData data) {
+    super(data, "RemoteDependency", "RemoteDependencyData");
     this.data = data;
   }
 
+  public void setId(String id) {
+    data.setId(truncateTelemetry(id, SanitizationHelper.MAX_ID_LENGTH, "RemoteDependency.id"));
+  }
+
   public void setName(String name) {
-    data.setName(truncateTelemetry(name, MAX_EVENT_NAME_LENGTH, "Event.name"));
+    data.setName(
+        truncateTelemetry(name, SanitizationHelper.MAX_NAME_LENGTH, "RemoteDependency.name"));
+  }
+
+  public void setResultCode(String resultCode) {
+    data.setResultCode(
+        truncateTelemetry(resultCode, MAX_RESULT_CODE_LENGTH, "RemoteDependency.resultCode"));
+  }
+
+  public void setData(String data) {
+    this.data.setData(truncateTelemetry(data, MAX_DATA_LENGTH, "RemoteDependency.data"));
+  }
+
+  public void setType(String type) {
+    data.setType(truncateTelemetry(type, MAX_DEPENDENCY_TYPE_LENGTH, "RemoteDependency.type"));
+  }
+
+  public void setTarget(String target) {
+    data.setTarget(truncateTelemetry(target, MAX_TARGET_NAME_LENGTH, "RemoteDependency.target"));
+  }
+
+  public void setDuration(String duration) {
+    data.setDuration(duration);
+  }
+
+  public void setSuccess(Boolean success) {
+    data.setSuccess(success);
   }
 
   public void addMeasurement(String key, Double value) {
