+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/legacysdk/BytecodeUtilImpl.java
@@ -28,15 +28,15 @@ import com.microsoft.applicationinsights.agent.internal.common.Strings;
 import com.microsoft.applicationinsights.agent.internal.exporter.models.ContextTagKeys;
 import com.microsoft.applicationinsights.agent.internal.exporter.models.DataPointType;
 import com.microsoft.applicationinsights.agent.internal.exporter.models.SeverityLevel;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.AbstractTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.EventTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.ExceptionTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.MessageTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.MetricPointBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.MetricTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.PageViewTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.RemoteDependencyTelemetryBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.RequestTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.builders.AbstractTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.builders.EventTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.builders.ExceptionTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.builders.MessageTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.builders.MetricPointBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.builders.MetricTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.builders.PageViewTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.builders.RemoteDependencyTelemetryBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.builders.RequestTelemetryBuilder;
 import com.microsoft.applicationinsights.agent.internal.init.AiOperationNameSpanProcessor;
 import com.microsoft.applicationinsights.agent.internal.legacyheaders.AiLegacyPropagator;
 import com.microsoft.applicationinsights.agent.internal.sampling.SamplingScoreGeneratorV2;
