+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/telemetry/TelemetryUtil.java
@@ -22,8 +22,8 @@
 package com.microsoft.applicationinsights.agent.internal.telemetry;
 
 import com.microsoft.applicationinsights.agent.internal.common.Strings;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.ExceptionDetailBuilder;
-import com.microsoft.applicationinsights.agent.internal.exporter.models.builders.StackFrameBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.builders.ExceptionDetailBuilder;
+import com.microsoft.applicationinsights.agent.internal.exporter.builders.StackFrameBuilder;
 import io.opentelemetry.api.trace.TraceState;
 import io.opentelemetry.instrumentation.api.cache.Cache;
 import java.util.ArrayList;
